+++ index.js
@@ -7,164 +7,54 @@
 //
 
 "use strict";
-import React, {Component} from "react";
-
-const OrientationNative = require("react-native").NativeModules.Orientation;
-const {NativeEventEmitter, Platform} = require("react-native");
-const LocalEventEmitter = new NativeEventEmitter(OrientationNative);
-
-var listeners = {};
-
-var id = 0;
-var META = "__listener_id";
-
-var locked = false;
-
-function getKey(listener) {
-    if (!listener.hasOwnProperty(META)) {
-        if (!Object.isExtensible(listener)) {
-            return "F";
-        }
-        Object.defineProperty(listener, META, {
-            value: "L" + ++id
-        });
-    }
-    return listener[META];
-}
 
 export default class Orientation {
-    static getOrientation = cb => {
-        OrientationNative.getOrientation(orientation => {
-            cb(orientation);
-        });
-    };
+  static getOrientation = cb => {
+    cb("UNKNOWN");
+  };
 
-    static getDeviceOrientation = cb => {
-        OrientationNative.getDeviceOrientation(deviceOrientation => {
-            cb(deviceOrientation);
-        });
-    };
+  static getDeviceOrientation = cb => {
+    cb("UNKNOWN");
+  };
 
-    static isLocked = () => {
-        return locked;
-    }
+  static isLocked = () => {
+    return false;
+  };
 
-    static lockToPortrait = () => {
-        locked = true;
-        OrientationNative.lockToPortrait();
-    };
+  static lockToPortrait = () => {};
 
-    static lockToPortraitUpsideDown = () => {
-        locked = true;
-        OrientationNative.lockToPortraitUpsideDown();
-    };
+  static lockToPortraitUpsideDown = () => {};
 
-    static lockToLandscape = () => {
-        locked = true;
-        OrientationNative.lockToLandscape();
-    };
+  static lockToLandscape = () => {};
 
-    static lockToLandscapeRight = () => {
-        locked = true;
-        OrientationNative.lockToLandscapeRight();
-    };
+  static lockToLandscapeRight = () => {};
 
-    static lockToLandscapeLeft = () => {
-        locked = true;
-        OrientationNative.lockToLandscapeLeft();
-    };
+  static lockToLandscapeLeft = () => {};
 
-    // OrientationMaskAllButUpsideDown
-    static lockToAllOrientationsButUpsideDown = () => {
-        locked = true;
-        if (Platform.OS === "ios") {
-            OrientationNative.lockToAllOrientationsButUpsideDown();
-        }
-    };
+  // OrientationMaskAllButUpsideDown
+  static lockToAllOrientationsButUpsideDown = () => {};
 
-    static unlockAllOrientations = () => {
-        locked = false;
-        OrientationNative.unlockAllOrientations();
-    };
+  static unlockAllOrientations = () => {};
 
-    static addOrientationListener = cb => {
-        var key = getKey(cb);
-        listeners[key] = LocalEventEmitter.addListener(
-            "orientationDidChange",
-            body => {
-                cb(body.orientation);
-            }
-        );
-    };
+  static addOrientationListener = cb => {};
 
-    static removeOrientationListener = cb => {
-        var key = getKey(cb);
-        if (!listeners[key]) {
-            return;
-        }
-        listeners[key].remove();
-        listeners[key] = null;
-    };
+  static removeOrientationListener = cb => {};
 
-    static addDeviceOrientationListener = cb => {
-        var key = getKey(cb);
-        listeners[key] = LocalEventEmitter.addListener(
-            "deviceOrientationDidChange",
-            body => {
-                cb(body.deviceOrientation);
-            }
-        );
-    };
+  static addDeviceOrientationListener = cb => {};
 
-    static removeDeviceOrientationListener = cb => {
-        var key = getKey(cb);
-        if (!listeners[key]) {
-            return;
-        }
-        listeners[key].remove();
-        listeners[key] = null;
-    };
+  static removeDeviceOrientationListener = cb => {};
 
-    static addLockListener = cb => {
-        var key = getKey(cb);
-        listeners[key] = LocalEventEmitter.addListener(
-            "lockDidChange",
-            body => {
-                cb(body.orientation);
-            }
-        );
-    };
+  static addLockListener = cb => {};
 
-    static removeLockListener = cb => {
-        var key = getKey(cb);
-        if (!listeners[key]) {
-            return;
-        }
-        listeners[key].remove();
-        listeners[key] = null;
-    };
+  static removeLockListener = cb => {};
 
-    static removeAllListeners = () => {
-        for(var key in listeners){
-            if (!listeners[key]) {
-                continue;
-            }
-            listeners[key].remove();
-            listeners[key] = null;
-        }
-    };
+  static removeAllListeners = () => {};
 
-    static getInitialOrientation = () => {
-        return OrientationNative.initialOrientation;
-    };
+  static getInitialOrientation = () => {
+    return "UNKNOWN";
+  };
 
-    static getAutoRotateState = cb => {
-        if (Platform.OS === "android") {
-            OrientationNative.getAutoRotateState(state => {
-                cb(state);
-            });
-        } else {
-            cb(true); // iOS not implement
-        }
-    };
+  static getAutoRotateState = cb => {
+    cb(true);
+  };
 }
